@@ -15,22 +15,23 @@ import java.util.Collections;
 @Service
 public class CustomUserDetailsService implements UserDetailsService {
 
-    @Autowired
-    private UserRepository userRepository;
+        @Autowired
+        private UserRepository userRepository;
 
-    @Override
-    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        User user = userRepository.findByUsername(username)
-                .orElseThrow(() -> new UsernameNotFoundException("User not found" + username));
-        return org.springframework.security.core.userdetails.User
-                .withUsername(user.getUsername())
-                .password(user.getPassword())
-                .authorities(Collections.singletonList(
-                        new SimpleGrantedAuthority(user.getRole().getAuthority())))
-                .accountExpired(!user.getAccountNonExpired())
-                .accountLocked(!user.getAccountNonLocked())
-                .credentialsExpired(!user.getCredentialsNonExpired())
-                .disabled(!user.getEnabled())
-                .build();
+        @Override
+        public UserDetails loadUserByUsername(String username)
+                        throws UsernameNotFoundException {
+                User user = userRepository.findByUsername(username)
+                                .orElseThrow(() -> new UsernameNotFoundException("User not found" + username));
+                return org.springframework.security.core.userdetails.User
+                                .withUsername(user.getUsername())
+                                .password(user.getPassword())
+                                .authorities(Collections.singletonList(
+                                                new SimpleGrantedAuthority(user.getRole().getAuthority())))
+                                .accountExpired(!user.getAccountNonExpired())
+                                .accountLocked(!user.getAccountNonLocked())
+                                .credentialsExpired(!user.getCredentialsNonExpired())
+                                .disabled(!user.getEnabled())
+                                .build();
         }
 }
