@@ -27,6 +27,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 @ExtendWith(MockitoExtension.class)
+@SuppressWarnings("null")
 public class ArtworkControllerTest {
 
     @Mock
@@ -87,7 +88,7 @@ public class ArtworkControllerTest {
         when(artworkService.updateArtwork(eq(1L), any(UpdateArtwork.class))).thenReturn(updatedResponse);
 
         mockMvc.perform(multipart("/api/v1/artworks/{id}", 1L)
-                        .file(image)
+                .file(image)
                 .param("name", "Moremi Updated")
                 .with(request -> {
                     request.setMethod("PATCH");
